+++ CHANGELOG.md
@@ -5,11 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Move player auto pause plugin to the tomate-gtk package
+
 ## 0.23.1
 
 ### Fixed
 
-- Script plugin name in the metadata file
+- The script plugin name in the metadata file
 
 ## 0.23.0